changed to double so that it gives more exact temperatures

+++ dataLog.cpp
@@ -23,7 +23,7 @@ void readval(atod sensor, string filename){
     time_t curr;
     curr=time(NULL);
     double result=sensor.voltage(1000);
-    int temp=((result*1.1)-0.5)*100;
+    double temp=((result*1.1)-0.5)*100;
     fstream file;
     file.open(filename.c_str(), fstream::in | fstream::out | fstream::app); //input and output streams, appending
     string new_line;
